# api/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "M9 — Reservas de viajes",
    "version": "0.1.0",
    "description": "Contrato de referencia de esta rama, basado en la guía de Bloque 3. Validar con Bloque 1. M2/M4/M5/M7 son stubs; X-Actor no autentica. Recordatorios, vencimiento automático y activación programada quedan fuera de TP1."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/reservas": {
      "get": {
        "operationId": "listarReservas",
        "summary": "Listar reservas",
        "tags": [
          "Reservas"
        ],
        "parameters": [
          {
            "name": "clienteId",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120
            }
          },
          {
            "name": "estado",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDIENTE",
                "CONFIRMADA",
                "CANCELADA",
                "ACTIVADA",
                "VENCIDA"
              ]
            }
          },
          {
            "name": "limite",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista ordenada por fechaHora e id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Reserva"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          }
        }
      },
      "post": {
        "operationId": "crearReserva",
        "summary": "Crear reserva confirmada",
        "tags": [
          "Reservas"
        ],
        "parameters": [
          {
            "name": "X-Actor",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120,
              "default": "cliente"
            },
            "description": "Trazabilidad de TP1. No implementa autenticación ni autorización."
          }
        ],
        "responses": {
          "201": {
            "description": "Reserva creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reserva"
                }
              }
            },
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "URL de la reserva creada"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NuevaReserva"
              }
            }
          }
        }
      }
    },
    "/reservas/{id}": {
      "get": {
        "operationId": "consultarReserva",
        "summary": "Consultar reserva",
        "tags": [
          "Reservas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reserva actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reserva"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          }
        }
      },
      "patch": {
        "operationId": "modificarReserva",
        "summary": "Modificar o reprogramar reserva editable",
        "tags": [
          "Reservas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Actor",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120,
              "default": "cliente"
            },
            "description": "Trazabilidad de TP1. No implementa autenticación ni autorización."
          }
        ],
        "responses": {
          "200": {
            "description": "Reserva actualizada; un PATCH sin cambios reales no agrega historial",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reserva"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificacionReserva"
              }
            }
          }
        },
        "description": "Solo PENDIENTE o CONFIRMADA y sin activación iniciada. Se rechaza cuando faltan menos del mínimo configurado; el límite exacto se admite. Cambio de recorrido o vehículo recalcula tarifa; fecha o zona solamente no lo hacen."
      }
    },
...
  },
  "components": {
    "schemas": {
      "Reserva": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "clienteId",
          "origen",
          "destino",
          "tipoVehiculo",
          "fechaHora",
          "zonaHoraria",
          "estado",
          "tarifaEstimada",
          "motivoCancelacion",
          "solicitudDespachoId",
          "creadoEn",
          "modificadoEn"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clienteId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "origen": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "destino": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "tipoVehiculo": {
            "type": "string",
            "enum": [
              "AUTO",
              "MOTO"
            ]
          },
          "fechaHora": {
            "type": "string",
            "format": "date-time",
            "description": "Instante ISO 8601 con Z u offset. Ventana configurable: 60 minutos a 30 días por defecto."
          },
          "zonaHoraria": {
            "type": "string",
            "maxLength": 100,
            "description": "Zona IANA de presentación. Puede combinarse con fechaHora en UTC."
          },
          "estado": {
            "type": "string",
            "enum": [
              "PENDIENTE",
              "CONFIRMADA",
              "CANCELADA",
              "ACTIVADA",
              "VENCIDA"
            ]
          },
          "tarifaEstimada": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "motivoCancelacion": {
            "type": "string",
            "nullable": true
          },
          "solicitudDespachoId": {
            "type": "string",
            "nullable": true
          },
          "creadoEn": {
            "type": "string",
            "format": "date-time"
          },
          "modificadoEn": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NuevaReserva": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "clienteId",
          "origen",
          "destino",
          "tipoVehiculo",
          "fechaHora",
          "zonaHoraria"
        ],
        "properties": {
          "clienteId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "origen": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "destino": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "tipoVehiculo": {
            "type": "string",
            "enum": [
              "AUTO",
              "MOTO"
            ]
          },
          "fechaHora": {
            "type": "string",
            "format": "date-time",
            "description": "Instante ISO 8601 con Z u offset. Ventana configurable: 60 minutos a 30 días por defecto."
          },
          "zonaHoraria": {
            "type": "string",
            "maxLength": 100,
            "description": "Zona IANA de presentación. Puede combinarse con fechaHora en UTC."
          }
        }
      },
      "ModificacionReserva": {
        "type": "object",
        "minProperties": 1,
        "additionalProperties": false,
        "properties": {
          "origen": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "destino": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "tipoVehiculo": {
            "type": "string",
            "enum": [
              "AUTO",
              "MOTO"
            ]
          },
          "fechaHora": {
            "type": "string",
            "format": "date-time",
            "description": "Instante ISO 8601 con Z u offset. Ventana configurable: 60 minutos a 30 días por defecto."
          },
          "zonaHoraria": {
            "type": "string",
            "maxLength": 100,
            "description": "Zona IANA de presentación. Puede combinarse con fechaHora en UTC."
          }
        }
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "codigo",
          "mensaje",
          "detalle"
        ],
        "properties": {
          "codigo": {
            "type": "string"
          },
          "mensaje": {
            "type": "string"
          },
          "detalle": {
            "nullable": true,
            "description": "Objeto o lista con detalles; null cuando no hay información adicional."
          }
        }
      },
...
    },
    "responses": {
      "Error503": {
        "description": "Dependencia indisponible o timeout",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error400": {
        "description": "Datos inválidos",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error404": {
        "description": "Reserva inexistente",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error409": {
        "description": "Estado incompatible, activación pendiente o conflicto concurrente",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error422": {
        "description": "Regla de negocio incumplida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error500": {
        "description": "Fallo interno sin exposición de datos sensibles",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error413": {
        "description": "Cuerpo demasiado grande",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}